requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0
pydantic==2.5.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
aiofiles==23.2.1
jinja2==3.1.2
pydantic-settings==2.10.1
aiohttp==3.12.15
python-multipart==0.0.20
pycryptodome==3.20.0